Put boat hit increment in its on function
Boat hits used to be incremented every time addHit() was
called, but this is not the correct use of this func.
Instead, addHit (now called addPrevHit) is supposed to
track previous shots made BY a player, while
incrementing of boat hits integer is only used when
there is a new hit made ON a player.