New branch to re-implement investigation functionality
Computer player needs the ability to intelligently
"investigate" when it makes a shot resulting in a hit. It
does this by procedurally checking coordinates adjacent to
the hit until it sinks the boat.

Current implementation of this functionality is overly
complex and should be re-done. This branch starts that
process, first by deleting all functions associate with the
current implementation.